+++ src/plots/ParallelCoordinates/QvisParallelCoordinatesPlotWindow.C
@@ -479,7 +479,7 @@ QvisParallelCoordinatesPlotWindow::UpdateWindow(bool doAll)
           case ParallelCoordinatesAttributes::ID_extentMaxima:
             axisTree->blockSignals(true);
             axisTree->clear();
-            for (int ax=0; ax<atts->GetExtentMinima().size(); ax++)
+            for (size_t ax=0; ax<atts->GetExtentMinima().size(); ax++)
             {
                 QString name, emin("min"), emax("max");
                 if (atts->GetExtentMinima()[ax] > -1e+37)
@@ -490,7 +490,7 @@ QvisParallelCoordinatesPlotWindow::UpdateWindow(bool doAll)
                     name = (atts->GetVisualAxisNames()[ax]).c_str();
                 else
                 {
-                    name.sprintf(" %02d",ax);
+                    name.sprintf(" %02ld",ax);
                     name = tr("Axis") + name;
                 }
                 QTreeWidgetItem *item =
@@ -961,11 +961,11 @@ QvisParallelCoordinatesPlotWindow::contextColorChanged(const QColor &color)
 void
 QvisParallelCoordinatesPlotWindow::resetAxisExtents()
 {
-    for (int i=0; i<atts->GetExtentMinima().size(); i++)
+    for (size_t i=0; i<atts->GetExtentMinima().size(); i++)
     {
         atts->GetExtentMinima()[i] = -1e+37;
     }
-    for (int i=0; i<atts->GetExtentMaxima().size(); i++)
+    for (size_t i=0; i<atts->GetExtentMaxima().size(); i++)
     {
         atts->GetExtentMaxima()[i] = +1e+37;
     }
@@ -1178,7 +1178,7 @@ QvisParallelCoordinatesPlotWindow::moveAxisDown()
     int naxes = axes.size();
  
     // can't move last axis down in list
-    if (index >= axes.size()-1)
+    if ((size_t)index >= axes.size()-1)
         return;
 
     // InsertAxis() will reorder axes already in the list, so we
